t0 unsigned int -> double. Add dummy musrview

+++ src/include/PRunBase.h
@@ -86,7 +86,7 @@ class PRunBase
 
     PRunData fData;         ///< data to be fitted
     double fTimeResolution; ///< time resolution
-    vector<int> fT0s;       ///< all t0's of a run! The derived classes will handle it
+    vector<double> fT0s;    ///< all t0's of a run! The derived classes will handle it
 
     virtual bool PrepareData() = 0; // pure virtual, i.e. needs to be implemented by the deriving class!!
